Implementation of services, adding the formResolver as well as some services for the registration form on the site.
