+++ config/services.yaml
@@ -4,7 +4,6 @@
 # Put parameters here that don't need to change on each machine where the app is deployed
 # https://symfony.com/doc/current/best_practices/configuration.html#application-related-configuration
 parameters:
-  picture_directory: '%kernel.project_dir%/public/uploads/profilepicture/'
   figure_image: '%kernel.project_dir%/public/uploads/figureimage/'
 
 
@@ -13,6 +12,8 @@ services:
     _defaults:
         autowire: true      # Automatically injects dependencies in your services.
         autoconfigure: true # Automatically registers your services as commands, event subscribers, etc.
+        bind:
+          $pictureLinkDirectory: '%kernel.project_dir%/public/uploads/profilepicture/'
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
