+++ src/Services/FormResolverPasswordRecovery.php
@@ -1,0 +1,16 @@
+<?php
+
+
+namespace App\Services;
+
+
+use Symfony\Component\Form\FormFactoryInterface;
+
+class FormResolverPasswordRecovery extends FormResolver
+{
+    public function __construct(FormFactoryInterface $formFactory)
+    {
+        parent::__construct($formFactory);
+    }
+
+}
